requirements.txt
python-dotenv==0.20.0
paho-mqtt==1.5.0